# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace Platform Api",
        "description": "Marketplace platform for develope",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@instafxgroup.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://shop/api/v1",
            "description": "Dev API server"
        }
    ],
    "paths": {
...
        "/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Category info",
                "description": "Category info",
                "operationId": "category.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category unfo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category",
                "description": "Update category",
                "operationId": "category.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Category data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "description": "Parent category ID",
                                        "type": "integer",
                                        "example": "1",
                                        "nullable": "true"
                                    },
                                    "title": {
                                        "description": "Title",
                                        "type": "string",
                                        "example": "Test",
                                        "nullable": "true"
                                    },
                                    "about": {
                                        "description": "Description",
                                        "type": "string",
                                        "example": "Some text",
                                        "nullable": "true"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "integer",
                                        "example": "0",
                                        "nullable": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category",
                "description": "Delete category",
                "operationId": "category.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destroy result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There are some active events in this stream OR Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Order info",
                "description": "Order info",
                "operationId": "order.show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order unfo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "$ref": "#/components/schemas/order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order",
                "description": "Update order",
                "operationId": "order.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Order data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "owner_id": {
                                        "description": "Ccategory ID",
                                        "type": "integer",
                                        "example": "1",
                                        "nullable": "true"
                                    },
                                    "about": {
                                        "description": "Description",
                                        "type": "string",
                                        "example": "Some text",
                                        "nullable": "true"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "integer",
                                        "example": "0",
                                        "nullable": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete order",
                "description": "Delete order",
                "operationId": "order.destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Destroy result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "There are some active events in this stream OR Error: Bad request. When required parameters were not supplied.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/error_response"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "error_response": {
                "title": "Error response",
                "properties": {
                    "error": {
                        "properties": {
                            "app": {
                                "type": "string",
                                "example": "Some Api"
                            },
                            "code": {
                                "type": "integer",
                                "example": "400"
                            },
                            "msg": {
                                "type": "string",
                                "example": "Bad request"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "user": {
                "title": "User data",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "verified": {
                        "type": "string",
                        "format": "date-time",
                        "example": "null",
                        "nullable": "true"
                    },
                    "created": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-04T16:33:11.000000Z"
                    },
                    "updated": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-04T16:33:11.000000Z"
                    }
                },
                "type": "object"
            },
            "category": {
                "title": "Category data",
                "properties": {
                    "category_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "parent": {
                        "type": "object",
                        "example": "category scheme",
                        "nullable": "true"
                    },
                    "title": {
                        "type": "integer",
                        "example": "Test"
                    },
                    "about": {
                        "type": "string",
                        "example": "Some text",
                        "nullable": "true"
                    },
                    "status": {
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "product": {
                "title": "Product data",
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "category": {
                        "$ref": "#/components/schemas/category"
                    },
                    "title": {
                        "type": "integer",
                        "example": "Test"
                    },
                    "about": {
                        "type": "string",
                        "example": "Some text",
                        "nullable": "true"
                    },
                    "unit": {
                        "type": "string",
                        "example": "kg"
                    },
                    "stock": {
                        "type": "numeric",
                        "format": "float",
                        "example": "kg"
                    },
                    "price": {
                        "type": "numeric",
                        "format": "float",
                        "example": "kg"
                    },
                    "options": {
                        "type": "string",
                        "example": "Some JSON data"
                    },
                    "status": {
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object"
            },
            "order": {
                "title": "Order data",
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "about": {
                        "type": "string",
                        "example": "Some text",
                        "nullable": "true"
                    },
                    "items": {
                        "type": "integer",
                        "example": "5"
                    },
                    "amount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "5000"
                    },
                    "discount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "350"
                    },
                    "total_amount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "4650"
                    },
                    "status": {
                        "type": "integer",
                        "example": "0"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/user"
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/order_log"
                        }
                    },
                    "positions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/order_position"
                        }
                    }
                },
                "type": "object"
            },
            "order_position": {
                "title": "Order position data",
                "properties": {
                    "record_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "order_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "product": {
                        "$ref": "#/components/schemas/product"
                    },
                    "price": {
                        "type": "numeric",
                        "format": "float",
                        "example": "5000"
                    },
                    "discount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "350"
                    },
                    "amount": {
                        "type": "numeric",
                        "format": "float",
                        "example": "4650"
                    }
                },
                "type": "object"
            },
            "order_log": {
                "title": "Order log data",
                "properties": {
                    "record_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "order_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "status": {
                        "type": "integer",
                        "example": "5"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Your token must be provided by Admin",
                "name": "Api Client",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}